No review needed.

+++ library/xgpl_src/IsoAgLib/util/impl/flexiblebytestrings.cpp
@@ -45,7 +45,7 @@ namespace __IsoAgLib {
     if ( apui8_srcStream != NULL )
       CNAMESPACE::memcpy( uint8, apui8_srcStream, 4 );
 #ifndef NO_8BIT_CHAR_TYPE
-    else uint32[0] = ~0UL;
+    else uint32[0] = ~uint32_t( 0UL );
 #else
     else
       for ( uint8_t index = 0; index < 4; index++ )
@@ -65,7 +65,7 @@ namespace __IsoAgLib {
 #ifdef HAS_64BIT_INT_TYPE
       uint64[0] = ~0ULL;
 #else
-      uint32[0] = uint32[1] = ~0UL;
+      uint32[0] = uint32[1] = ~uint32_t( 0UL );
 #endif
 #endif
     }
